pyproject.toml
[tool.poetry]
name = "time-series-management-system"
version = "0.1.0"
description = ""
authors = ["Mikiyas <mikiyas.m.degefu@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
django = "^4.2.6"
django-import-export = "4.1.1"
django-fontawesome-5 = "^1.0.18"
djangorestframework = "3.15.2"
pillow = "<11.0"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
